Added more feature to the cart page
